pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.102.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
